requirements.txt
Flask
flask-cors
transformers
requests
python-dotenv
pyngrok